Merges data into an email and send.

Merges columns from a data frame into a markdown document using the glue::glue_data() function. The markdown can contain a yaml header for subject and cc line.

mail_merge(
  data,
  message,
  to_col = "email",
  preview = TRUE,
  draft = TRUE,
  confirm = FALSE,
  sleep_preview = 1,
  sleep_send = 0.1
)

Arguments

data
A data frame or tibble with all the columns that should be glued into the message. Substitution is performed using glue::glue_data()`

message
A list with components yaml and body. You can use mm_read_message() or mm_read_message_googledoc() to construct a message in this format.

to_col
The name of the column in data that contains the email address to send the message to.

preview
If TRUE displays message in viewer without sending mail.

draft
if TRUE sends a draft, otherwise sends the real email

confirm
If TRUE sends email without additional confirmation. If FALSE asks for confirmation before sending.

sleep_preview
If draft == TRUE the number of seconds to sleep between each preview.

sleep_send
If draft == FALSE the number of seconds to sleep between each email send (to prevent gmail API 500 errors).

Value

invisibly returns the input data set. This means you can potentially pipe multiple operations together.

Examples

## ---- input-data --------------------------------------------------------

dat <-  data.frame(
  email      = c("friend@example.com", "foe@example.com"),
  first_name = c("friend", "foe"),
  thing      = c("something good", "something bad"),
  stringsAsFactors = FALSE
)

## ---- markdown-message --------------------------------------------------

msg <- '
---
subject: "**Hello, {first_name}**"
---

Hi, **{first_name}**

I am writing to tell you about **{thing}**.

{if (first_name == "friend") "Regards" else ""}

Me
'

## ---- mail-merge --------------------------------------------------------

dat %>%
  mail_merge(msg)

#> Sent preview to viewer

if (interactive()) {
  dat %>%
    mail_merge(msg) %>%
    print()
}

if (interactive()) {
  dat %>%
    mail_merge(msg) %>%
    preview_mailmerge()
}